refactor: abstracting datastore configuration retrieval
